moved cpr to own file
