+++ src/adsb/aircraft.rs
@@ -6,12 +6,9 @@ use chrono::{Local};
 
 use crate::adsb::msgs::{AdsbMsgType, CprFormat};
 use crate::adsb;
+use crate::adsb::cpr::{calculate_geographic_position, GeographicPosition};
 
-#[derive(Debug, Clone)]
-pub struct GeographicPosition {
-    pub latitude: f64,
-    pub longitude: f64,
-}
+
 
 /// Holder for aircraft information that has been received from adsb
 #[derive(Debug)]
@@ -101,108 +98,6 @@ impl Aircraft {
 }
 
 
-const NUM_ZONES: f64 = 15.0;
-
-/// Convert a CPR value to a float
-fn convert_cpr_to_float(cpr: u32) -> f64 {
-    const CPR_TO_FLOAT: f64 = 131072.0; // 2^17
-    (cpr as f64) / CPR_TO_FLOAT
-}
-
-/// Find the number of longitude zones for a given latitude
-/// 
-/// `lat` - the latitude in degrees
-/// 
-/// returns the number of longitude zones
-fn calc_num_zones(lat: f64) -> u32 {
-    if lat == 0.0 {
-        return 59; // Special case for equator
-    } else if lat == 87.0 || lat == -87.0 {
-        return 2; // Special case for poles
-    } else if lat < -87.0 || lat > 87.0 {
-        return 1; // Invalid latitude
-    }
-
-    let pi = std::f64::consts::PI;
-    let int1 = 1.0 - (pi/(2.0*NUM_ZONES)).cos();
-    let int2 = (pi/180.0 * lat).cos();
-    let int3 = (2.0 * pi) / (1.0 - (int1/(int2*int2))).acos();
-
-    int3.floor() as u32
-}
-
-/// Calculate the latitude from the even and odd CPR latitudes
-fn calculate_latitude(even_cpr_lat: u32, odd_cpr_lat: u32, first: CprFormat) -> f64{
-    
-    const EVEN_LAT_DIVISIONS: f64 = 360.0 / (4.0 * NUM_ZONES);
-    const ODD_LAT_DIVISIONS: f64 = 360.0 / (4.0 * NUM_ZONES-1.0);
-
-    let even_cpr_lat = convert_cpr_to_float(even_cpr_lat);
-    let odd_cpr_lat = convert_cpr_to_float(odd_cpr_lat);
-
-    let latitude_index: f64 = (59.0*even_cpr_lat - 60.0 * odd_cpr_lat + 0.5).floor();
-
-    let even_latitude = EVEN_LAT_DIVISIONS * (latitude_index % 60.0 + even_cpr_lat);
-    let odd_latitude = ODD_LAT_DIVISIONS * (latitude_index % 59.0 + odd_cpr_lat);
-
-    let mut latitude = match first {
-        // Use the newest format to determine the latitude
-        CprFormat::Even => odd_latitude,
-        CprFormat::Odd => even_latitude,
-    };
-
-    if latitude > 270.0 {
-        latitude -= 360.0;
-    }
-
-    latitude
-    
-}
-
-fn calculate_longitude(even_cpr_long: u32, odd_cpr_long: u32, latitude: f64, first: CprFormat) -> f64 {
-    let lon_cpr_e = convert_cpr_to_float(even_cpr_long);
-    let lon_cpr_o = convert_cpr_to_float(odd_cpr_long);
-
-    let num_zones: f64;
-
-    match first {
-        CprFormat::Even => {
-            // Later is odd
-            num_zones = (calc_num_zones(latitude)-1).max(1) as f64
-        },
-        CprFormat::Odd => {
-            // Later is even
-            num_zones = (calc_num_zones(latitude)).max(1) as f64
-        },
-        
-    }
-
-    
-    let divisions = 360.0 / num_zones;
-    let m = (lon_cpr_e * (num_zones - 1.0) - lon_cpr_o * num_zones + 0.5).floor();
-    let mut longitude = divisions * ((m % num_zones) + lon_cpr_o);
-
-    if longitude > 180.0 {
-        longitude -= 360.0;
-    }
-
-    longitude
-}
-
-/// Calculate the geographic position from the even and odd CPR positions
-/// 
-/// `even_cpr_lat_long` - the latitude and longitude from the even CPR packet
-/// `odd_cpr_lat_long` - the latitude and longitude from the odd CPR packet
-/// `first` - the first CPR format (even or odd)
-/// 
-/// returns a GeographicPosition
-fn calculate_geographic_position(even_cpr_lat_long: (u32, u32), odd_cpr_lat_long: (u32, u32), first: CprFormat) -> GeographicPosition {
-    let latitude = calculate_latitude(even_cpr_lat_long.0, odd_cpr_lat_long.0, first);
-    let longitude = calculate_longitude(even_cpr_lat_long.1, odd_cpr_lat_long.1, latitude, first);
-
-
-    GeographicPosition { latitude: latitude, longitude: longitude }
-}
 
 mod tests {
     #[allow(unused_imports)]
@@ -236,42 +131,6 @@ mod tests {
         let packet = AdsbPacket::_new_from_string(String::from_str("8d7c6b30581304f388bb4455896f").unwrap());
         aircraft.handle_packet(packet);
         assert_eq!(aircraft.get_altitude_ft(), 2600);
-    }
-
-    #[test]
-    fn test_latitude_calculation() {
-        let even_cpr_lat = 93000; // Example even CPR latitude
-        let odd_cpr_lat = 74158; // Example odd CPR latitude
-        let first = CprFormat::Odd;
-
-        let latitude = calculate_latitude(even_cpr_lat, odd_cpr_lat, first);
-        assert!((latitude - 52.25720).abs() < 0.0001); // Adjust expected value based on actual calculation
-    }
-
-    #[test]
-    fn test_zone_calcuation() {
-        assert_eq!(calc_num_zones(0.0), 59);
-
-        assert_eq!(calc_num_zones(87.0), 2);
-
-        assert_eq!(calc_num_zones(-87.0), 2);
-
-        assert_eq!(calc_num_zones(90.0), 1);
-
-        assert_eq!(calc_num_zones(-90.0), 1);
-
-        assert_eq!(calc_num_zones(10.0), 7);
-    }
-
-    #[test]
-    fn test_longitude_calculation() {
-        let even_cpr_long = 51372; // Example even CPR longitude
-        let odd_cpr_long = 50194; // Example odd CPR longitude
-        let latitude = 52.25720214843750; // Example latitude
-        let first = CprFormat::Odd;
-
-        let longitude = calculate_longitude(even_cpr_long, odd_cpr_long, latitude, first);
-        assert!((longitude -  3.829498291015625).abs() < 0.0001); // Adjust expected value based on actual calculation
     }
 
     #[test]
